+++ src/demo/app-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule }  from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 import { EmptyStateExampleComponent } from '../app/empty-state/examples/empty-state-example.component';
-import { FilterExampleComponent } from '../app/filters/examples/filter-example.component';
+import { FilterExampleComponent } from '../app/filter/examples/filter-example.component';
 import { ListViewExampleComponent } from '../app/list-view/examples/list-view-example.component';
 import { RemainingCharsExampleComponent } from '../app/remaining-chars/examples/remaining-chars-example.component';
 import { SampleExampleComponent } from '../app/sample/examples/sample-example.component';
