+++ src/demo/app.module.ts
@@ -12,7 +12,7 @@ import { AppRoutingModule } from './app-routing.module';
 // Main areas
 //  import example modules
 import { EmptyStateExampleModule } from '../app/empty-state/examples/empty-state-example.module';
-import { FilterExampleModule } from '../app/filters/examples/filter-example.module';
+import { FilterExampleModule } from '../app/filter/examples/filter-example.module';
 import { NotificationExampleModule } from '../app/notification/examples/notification-example.module';
 import { ListViewExampleModule } from '../app/list-view/examples/list-view-example.module';
 import { RemainingCharsExampleModule } from '../app/remaining-chars/examples/remaining-chars-example.module';
